Module split
The code is easier to browse if we split the state and the existential
code into sub-modules.
